Move plugin headers to plugins directory
